electron-app/python-bridge/requirements.txt
# Vinted UK Sniper — Python Bridge
# Stealth HTTP client for bypassing Cloudflare/Datadome

fastapi>=0.109.0
uvicorn[standard]>=0.27.0
curl_cffi>=0.9.0
Pillow>=10.0.0
piexif>=1.1.3
imagehash>=4.3.0
numpy>=1.24.0
python-multipart>=0.0.6

# Item Intelligence — AI pipeline dependencies
pydantic>=2.6.0
httpx>=0.27.0
sse-starlette>=2.0.0
